Feat: Listas de adjacência com objetos Aresta

+++ objetos/vertice.py
@@ -3,5 +3,5 @@ import numpy as np
 
 class Vertice:
     def __init__(self):
-        self.lista_adjacencia = np.array([], dtype="int")
+        self.lista_adjacencia = np.array([])
         self.dado = 0
